+++ inst/scripts/run_oncoscan_workflow.R
@@ -15,7 +15,7 @@ suppressPackageStartupMessages(library(oncoscanR))
 #' Retrieve arguments from command line.
 #' Expects that the first argument is the ChAS file.
 args <- commandArgs(TRUE)
-if(length(args) != 2){
+if(length(args) != 1){
   stop("The first argument has to be the ChAS file ")
 }