+++ tests/testthat/test-input.R
@@ -21,7 +21,8 @@ test_that("loading ChAS file works", {
     }
   }
 
-  expect_equal(length(segs), 15)  # 14 segments in file but one is covering two arms
+  expect_equal(length(segs), 15)  
+  # 14 segments in file but one is covering two arms
   expect_equal(found, 2, info = print(found))
 
 })
@@ -38,13 +39,15 @@ test_that("loading large ChAS file", {
   segs.filename <- system.file("extdata", "LST_gene_list_full_location.txt",
                                package = "oncoscanR")
   segs <- load_chas(segs.filename, kit.coverage = oncoscan_na33.cov)
-  expect_true(length(segs)>100) # Mostly testing that it does not fail nor return zero segments
+  expect_true(length(segs)>100) 
+  # Mostly testing that it does not fail nor return zero segments
 })
 
 
 test_that("Loading ChAS with missing 'Full Location' column fails",{
   segs.filename <- "../testdata/chas_example-noFullLocation.txt"
-  expect_error(suppressWarnings(load_chas(segs.filename, kit.coverage = oncoscan_na33.cov)),
+  expect_error(suppressWarnings(load_chas(segs.filename, 
+                                          kit.coverage = oncoscan_na33.cov)),
                'Parsing ChAS file failed')
 })
 
@@ -69,12 +72,14 @@ test_that("loading ChAS with 'chr' name scheme works", {
     }
   }
 
-  expect_equal(c(length(segs), found), c(15, 2)) # 14 segments in file but one is covering two arms
+  expect_equal(c(length(segs), found), c(15, 2)) 
+  # 14 segments in file but one is covering two arms
 })
 
 test_that("Loading ChAS with empty file works",{
   segs.filename <- "../testdata/chas_example-empty.txt"
-  expect_warning(segs <- load_chas(segs.filename, kit.coverage = oncoscan_na33.cov),
+  expect_warning(segs <- load_chas(segs.filename, 
+                                   kit.coverage = oncoscan_na33.cov),
                  'No segments loaded')
   expect_equal(length(segs), 0)
 })
@@ -104,5 +109,13 @@ test_that("Loading ASCAT file works", {
     
     expect_equal(length(segs), 2)  # 2 segments in file 
     expect_equal(found, 2, info = print(found))
+})
+
+test_that("Loading ASCAT with empty file works",{
+    segs.filename <- "../testdata/ascat_example-empty.txt"
+    expect_warning(segs <- load_ascat(segs.filename, 
+                                      kit.coverage = oncoscan_na33.cov),
+                   'No segments loaded')
+    expect_equal(length(segs), 0)
 })