+++ tests/testthat/test-workflow.R
@@ -6,7 +6,7 @@ test_that("Oncoscan workflow works - LST", {
   armlevel.fn <- "../testdata/LST_gene_list_full_location.armlevel_scna.csv"
 
   # Run standard workflow from package
-  dat <- workflow_oncoscan.run(chas.fn)
+  dat <- workflow_oncoscan.chas(chas.fn)
 
   tests <- armlevel.test(dat[['armlevel']], armlevel.fn)
   expect_true(sum(tests)==4)
@@ -25,7 +25,7 @@ test_that("Oncoscan workflow works - triploide", {
   armlevel.fn <- "../testdata/triploide_gene_list_full_location.armlevel_scna.csv"
 
   # Run standard workflow from package
-  dat <- workflow_oncoscan.run(chas.fn)
+  dat <- workflow_oncoscan.chas(chas.fn)
 
   tests <- armlevel.test(dat[['armlevel']], armlevel.fn)
   expect_true(sum(tests)==4)
@@ -44,7 +44,7 @@ test_that("Oncoscan workflow works - TDplus", {
   armlevel.fn <- "../testdata/TDplus_gene_list_full_location.armlevel_scna.csv"
 
   # Run standard workflow from package
-  dat <- workflow_oncoscan.run(chas.fn)
+  dat <- workflow_oncoscan.chas(chas.fn)
 
   tests <- armlevel.test(dat[['armlevel']], armlevel.fn)
   expect_true(sum(tests)==4)
@@ -57,7 +57,7 @@ test_that("Oncoscan workflow works - TDplus", {
 
 test_that("Oncoscan workflow returns the correct structure", {
   segs.filename <- system.file("extdata", "chas_example.txt", package = "oncoscanR")
-  dat <- workflow_oncoscan.run(segs.filename)
+  dat <- workflow_oncoscan.chas(segs.filename)
 
   armlevel.nametest <- identical(sort(names(dat[['armlevel']])), sort(c('AMP', 'GAIN', 'LOH', 'LOSS')))
   scores.nametest <- identical(sort(names(dat[['scores']])), sort(c("avgCN", "HRD", "TDplus")))
@@ -71,7 +71,7 @@ test_that("Oncoscan workflow works with an empty file", {
   segs.filename <- "../testdata/chas_example-empty.txt"
   armlevel.fn <- "../testdata/chas_example-empty.armlevel_scna.csv"
 
-  expect_warning(dat <- workflow_oncoscan.run(segs.filename))
+  expect_warning(dat <- workflow_oncoscan.chas(segs.filename))
 
   # Test arm-level detection
   tests <- armlevel.test(dat[['armlevel']], armlevel.fn)
@@ -82,4 +82,40 @@ test_that("Oncoscan workflow works with an empty file", {
   expect_equal(as.character(scores['HRD']), "Negative (no tumor?), nLST=0")
   expect_equal(as.character(scores['avgCN']), '2')
   expect_equal(as.character(scores['TDplus']), '0')
+})
+
+test_that("Oncoscan ASCAT workflow works", {
+    #skip_on_cran()
+    
+    ascat.fn <- system.file("extdata", "ascat_example.txt",
+                           package = "oncoscanR")
+    armlevel.fn <- "../testdata/ascat_example.armlevel_scna.csv"
+    
+    # Run standard workflow from package
+    dat <- workflow_oncoscan.ascat(ascat.fn)
+    
+    tests <- armlevel.test(dat[['armlevel']], armlevel.fn)
+    expect_true(sum(tests)==4)
+    
+    scores <- dat[['scores']]
+    expect_equal(as.character(scores['HRD']), "Negative, nLST=0")
+    expect_equal(as.character(scores['avgCN']), '2.05')
+    expect_equal(as.character(scores['TDplus']), '0')
+})
+
+test_that("ASCAT workflow works with an empty file", {
+    segs.filename <- "../testdata/ascat_example-empty.txt"
+    armlevel.fn <- "../testdata/chas_example-empty.armlevel_scna.csv"
+    
+    expect_warning(dat <- workflow_oncoscan.ascat(segs.filename))
+    
+    # Test arm-level detection
+    tests <- armlevel.test(dat[['armlevel']], armlevel.fn)
+    expect_true(sum(tests)==4)
+    
+    # Test scores
+    scores <- unlist(dat[['scores']])
+    expect_equal(as.character(scores['HRD']), "Negative (no tumor?), nLST=0")
+    expect_equal(as.character(scores['avgCN']), '2')
+    expect_equal(as.character(scores['TDplus']), '0')
 })