+++ NEWS.md
@@ -5,6 +5,7 @@
 `workflow_oncoscan.chas` for consistency with the ASCAT function
 - Bug fixes in adjust_loh (crashed with segments of length 1 or if no LOH 
 segments where present)
+- Bug fix in script `run_oncoscan_workflow.R`
 
 ## What's new in v1.1.0